Introduce ClientObserveRelation.
